Task #4585: reenable tBeamFormerSubbandProcProcessSb: fix calculation of threadDim.z, add a few Device getAttribute functions for device limit detection, turn maxNrThreads type into unsigned (enough), turn a few NR_POLARIZATION macros into var reads

+++ branches/Cobalt-Task4585/RTCP/Cobalt/GPUProc/src/cuda/Kernels/Kernel.h
@@ -85,7 +85,7 @@ namespace LOFAR
       // Explicit destructor, because the implicitly generated one is public.
       ~Kernel();
       
-      const size_t maxThreadsPerBlock;
+      const unsigned maxThreadsPerBlock;
       gpu::Grid globalWorkSize;
       gpu::Block localWorkSize;
       size_t nrOperations, nrBytesRead, nrBytesWritten;
